Extract recipe runtime setup helper
