+++ packages/cli/src/commands/recipe-run.ts
@@ -1,18 +1,17 @@
 import { createHash } from "node:crypto"
-import { cp, mkdir, mkdtemp, readdir, readFile, rm, stat, writeFile } from "node:fs/promises"
-import { tmpdir } from "node:os"
+import { mkdir, readdir, readFile, stat, writeFile } from "node:fs/promises"
 import { basename, dirname, join, resolve } from "node:path"
-import { DEFAULT_CAPTURED_ARTIFACT_MAX_BYTES, DEFAULT_WORDPRESS_VERSION, STRUCTURED_ARTIFACT_SCHEMA, TYPED_ARTIFACT_INDEX_SCHEMA, artifactBundleRunRef, artifactFileDigest, assertFixtureImportDeterministicIdsSupported, createRuntime, fixtureImportDeterministicIdPlan, normalizeRecipeRunSummary, normalizeRuntimeEnvRecord, parseCommandOptions, redactJsonValue, resolveSecretEnvNames, workspaceRecipeRuntimeCollectedArtifacts, type ArtifactBundle, type ArtifactManifestFile, type ExecutionResult, type Runtime, type RuntimeAssetSpec, type RuntimePreviewSpec, type RuntimeRunRecord, type RuntimeRunRegistry, type TypedArtifactIndex, type TypedArtifactRef, type WorkspaceRecipe, type WorkspaceRecipeComponentManifest, type WorkspaceRecipeDeclaredArtifact, type WorkspaceRecipeExtraPlugin, type WorkspaceRecipeFixtureDatabase, type WorkspaceRecipeMount, type WorkspaceRecipePluginRuntimeHealthProbe, type WorkspaceRecipeProbe, type WorkspaceRecipeSiteSeed, type WorkspaceRecipeTypedArtifact } from "@automattic/wp-codebox-core"
+import { DEFAULT_CAPTURED_ARTIFACT_MAX_BYTES, DEFAULT_WORDPRESS_VERSION, STRUCTURED_ARTIFACT_SCHEMA, TYPED_ARTIFACT_INDEX_SCHEMA, artifactBundleRunRef, artifactFileDigest, assertFixtureImportDeterministicIdsSupported, createRuntime, fixtureImportDeterministicIdPlan, normalizeRecipeRunSummary, normalizeRuntimeEnvRecord, parseCommandOptions, redactJsonValue, resolveSecretEnvNames, workspaceRecipeRuntimeCollectedArtifacts, type ArtifactBundle, type ArtifactManifestFile, type ExecutionResult, type Runtime, type RuntimeAssetSpec, type RuntimePreviewSpec, type RuntimeRunRecord, type RuntimeRunRegistry, type TypedArtifactIndex, type TypedArtifactRef, type WorkspaceRecipe, type WorkspaceRecipeComponentManifest, type WorkspaceRecipeDeclaredArtifact, type WorkspaceRecipeExtraPlugin, type WorkspaceRecipeFixtureDatabase, type WorkspaceRecipeProbe, type WorkspaceRecipeSiteSeed, type WorkspaceRecipeTypedArtifact } from "@automattic/wp-codebox-core"
 import { stripUndefined } from "@automattic/wp-codebox-core/internals"
 import { recipeExecutionSpec, sandboxWorkspaceContract } from "../agent-sandbox.js"
 import { executeAgentFanoutFromArgs } from "../agent-fanout.js"
 import { captureStdout, printRecipeHumanOutput, printRecipeValidateHumanOutput, serializeError } from "../output.js"
 import { parsePreviewBind, parsePreviewHoldSeconds, parsePreviewPort, parsePreviewPublicUrl } from "../preview-options.js"
-import { dryRunRecipe, planWorkspaceRecipe, pluginRuntimeHealthProbeStepIndex, pluginRuntimeSetupStepIndex, recipeDryRunSiteSeeds, siteSeedScopesAreBounded } from "../recipe-dry-run.js"
+import { dryRunRecipe, planWorkspaceRecipe, recipeDryRunSiteSeeds, siteSeedScopesAreBounded } from "../recipe-dry-run.js"
 import { appendRecipeRuntimeEvidence, appendRecipeRuntimeEvidenceFiles, collectAndFinalizeFailedRecipeArtifacts, collectRecipeRuntimeArtifacts, finalizeAgentSandboxEvidence, finalizeRecipeArtifactEvidence, recipeAgentResultFailure, recipeAgentResultOutput, recipeAgentTaskResultOutput, recipeArtifactEvidenceFailure, recipeCompletionOutcomeOutput, recipeReplayStatusOutput, recipeTerminalResultOutput, recipeVerifyStepFailure } from "../recipe-evidence.js"
-import { prepareRecipeRuntimeBackendPackage, type PreparedRuntimeBackendPackage } from "../recipe-backend-package.js"
-import { cleanupRecipePreparedSources, installMuPluginsCode, prepareRecipeDependencyOverlays, prepareRecipeExtraPlugins, prepareRecipeRuntimeOverlays, prepareRecipeStagedFiles, prepareRecipeWorkspaces, recipeBlueprintWithBootActivePlugins, recipeExtraPlugins, recipeMountType, type PreparedDependencyOverlay, type PreparedExtraPlugin, type PreparedRuntimeOverlay, type PreparedStagedFile, type PreparedWorkspaceMount } from "../recipe-sources.js"
-import { loadWorkspaceRecipe, pluginRuntimeHealthProbeStep, recipePolicy, recipeWorkflowSteps, validateWorkspaceRecipe, type RecipeWorkflowPhase } from "../recipe-validation.js"
+import type { PreparedRuntimeBackendPackage } from "../recipe-backend-package.js"
+import { cleanupRecipePreparedSources, recipeBlueprintWithBootActivePlugins, recipeExtraPlugins, type PreparedDependencyOverlay, type PreparedExtraPlugin, type PreparedRuntimeOverlay, type PreparedStagedFile, type PreparedWorkspaceMount } from "../recipe-sources.js"
+import { loadWorkspaceRecipe, recipePolicy, recipeWorkflowSteps, validateWorkspaceRecipe, type RecipeWorkflowPhase } from "../recipe-validation.js"
 import { resolveCliRuntimeBackend } from "../runtime-backends.js"
 import { previewSpec, releaseRuntime, runtimeMetadata, type RunOutput } from "../runtime-command-wrappers.js"
 import { artifactManifestFilesByPath, parseBenchResults, writeBenchmarkArtifactEvidence } from "./recipe-run-benchmark-artifacts.js"
@@ -21,6 +20,7 @@ import { RecipeRunPhaseExecutor } from "./recipe-run-phase-executor.js"
 import { createRecipeInterruptionController, interruptedRecipeOutput, markRecipeArtifactsFinalized, recipeInterruptionSerializedError } from "./recipe-run-interruption.js"
 import { bestEffortTimeout, exitAfterPlaygroundCliBootFailure, exitAfterRecipeRunTimeout, exitAfterTerminalRecipePhaseFailure, printJsonFailureDiagnostic, recipeRunFailureStatus, RecipeDeclaredArtifactFailureError, RecipeProbeFailureError, RecipeRunTimeoutError, RecipeRuntimeCreateError, serializeRecipeRunError, writeRecipeJsonOutput } from "./recipe-run-output.js"
 import { RecipePhaseError } from "./recipe-run-phases.js"
+import { applyRecipeRuntimeSetup, cleanupInputMountBaselines, prepareRecipeRuntimeSetup, recipeRunDependencyOverlay, recipeRunExtraPlugin, recipeRunStagedFile } from "./recipe-runtime-setup.js"
 import type { RecipeAdvisoryFailure, RecipeBrowserEvidence, RecipeBrowserEvidenceFileRef, RecipeDiagnosticArtifactRef, RecipeExecutionResult, RecipeInterruptionController, RecipePhaseEvidence, RecipePhaseName, RecipePhpWasmRuntimeDiagnostic, RecipeRunCommandOutput, RecipeRunComponentContract, RecipeRunDeclaredArtifact, RecipeRunFixtureDatabase, RecipeRunOptions, RecipeRunOutput, RecipeRunProbe, RecipeRunSiteSeed, RecipeRunStagedFile, RecipeRuntimeDiagnostic, RecipeValidateOptions, RecipeValidateOutput } from "./recipe-run-types.js"
 
 const DEFAULT_RECIPE_RUN_TIMEOUT_MS = 25 * 60 * 1000
@@ -113,7 +113,7 @@ async function runRecipe(options: RecipeRunOptions, interruption?: RecipeInterru
   let dependencyOverlays: PreparedDependencyOverlay[] = []
   let stagedFiles: PreparedStagedFile[] = []
   let overlays: PreparedRuntimeOverlay[] = []
-  const inputMountBaselinePaths: string[] = []
+  let inputMountBaselinePaths: string[] = []
   let backendPackage: PreparedRuntimeBackendPackage | undefined
   let runtime: Awaited<ReturnType<typeof createRuntime>> | undefined
   const executions: RecipeExecutionResult[] = []
@@ -139,12 +139,8 @@ async function runRecipe(options: RecipeRunOptions, interruption?: RecipeInterru
   const awaitRecipe = <T>(operation: string, promiseOrFactory: Promise<T> | (() => Promise<T>), timeoutMs?: number): Promise<T> => phaseExecutor.operation(operation, promiseOrFactory, timeoutMs)
 
   try {
-    workspaceMounts = await prepareRecipeWorkspaces(recipe, recipeDirectory)
-    extraPlugins = await prepareRecipeExtraPlugins(recipe, recipeDirectory)
-    dependencyOverlays = await prepareRecipeDependencyOverlays(recipe, recipeDirectory, extraPlugins)
-    stagedFiles = await prepareRecipeStagedFiles(recipe, recipeDirectory)
-    overlays = await prepareRecipeRuntimeOverlaysForRun(recipe, recipeDirectory)
-    backendPackage = await prepareRecipeRuntimeBackendPackage(recipe, recipeDirectory, plan.runtime.backend)
+    const preparedRuntimeSetup = await prepareRecipeRuntimeSetup(recipe, recipeDirectory, plan.runtime.backend)
+    ;({ workspaceMounts, extraPlugins, dependencyOverlays, stagedFiles, overlays, inputMountBaselinePaths, backendPackage } = preparedRuntimeSetup)
     interruption?.throwIfInterrupted()
 
     runRecord = await runRegistry.update(runRecord.runId, { status: "booting" })
@@ -227,125 +223,7 @@ async function runRecipe(options: RecipeRunOptions, interruption?: RecipeInterru
     })
     interruption?.throwIfInterrupted()
 
-    for (const [index, mount] of (recipe.runtime?.stack?.mounts ?? []).entries()) {
-      const source = resolve(recipeDirectory, mount.source)
-      await awaitRecipe(`runtime.stack.mount[${index}]`, runtime.mount({
-        type: await recipeMountType(source, mount.type),
-        source,
-        target: mount.target,
-        mode: mount.mode ?? "readonly",
-        metadata: {
-          kind: "runtime-stack-mount",
-          index,
-          ...(mount.metadata ?? {}),
-        },
-      }))
-      interruption?.throwIfInterrupted()
-    }
-
-    for (const overlay of overlays) {
-      await awaitRecipe(`runtime.overlay.mount:${overlay.target}`, runtime.mount({
-        type: overlay.type,
-        source: overlay.source,
-        target: overlay.target,
-        mode: overlay.mode,
-        metadata: overlay.metadata,
-      }))
-      interruption?.throwIfInterrupted()
-    }
-
-    for (const workspace of workspaceMounts) {
-      await awaitRecipe(`workspace.mount:${workspace.target}`, runtime.mount({
-        type: "directory",
-        source: workspace.source,
-        target: workspace.target,
-        mode: workspace.mode,
-        metadata: workspace.metadata,
-      }))
-      interruption?.throwIfInterrupted()
-    }
-
-    await phaseTracker.run("mount_plugins", phasePluginMountData(extraPlugins), async () => {
-      for (const plugin of extraPlugins) {
-        await awaitRecipe(`extra-plugin.mount:${plugin.slug}`, runtime!.mount({
-          type: "directory",
-          source: plugin.source,
-          target: plugin.target,
-          mode: "readonly",
-          metadata: {
-            kind: "extra-plugin",
-            slug: plugin.slug,
-            source: plugin.provenance,
-          },
-        }))
-        interruption?.throwIfInterrupted()
-      }
-    })
-
-    for (const overlay of dependencyOverlays) {
-      await awaitRecipe(`dependency-overlay.mount:${overlay.package}`, runtime.mount({
-        type: overlay.type,
-        source: overlay.source,
-        target: overlay.target,
-        mode: overlay.mode,
-        metadata: overlay.metadata,
-      }))
-      interruption?.throwIfInterrupted()
-    }
-
-    for (const mount of recipe.inputs?.mounts ?? []) {
-      const source = resolve(recipeDirectory, mount.source)
-      const metadata = await inputMountMetadataWithBaseline(source, mount, inputMountBaselinePaths)
-      await awaitRecipe(`input.mount:${mount.target}`, runtime.mount({
-        type: await recipeMountType(source, mount.type),
-        source,
-        target: mount.target,
-        mode: mount.mode ?? "readwrite",
-        metadata,
-      }))
-      interruption?.throwIfInterrupted()
-    }
-
-    for (const stagedFile of stagedFiles) {
-      await awaitRecipe(`staged-file.mount:${stagedFile.target}`, runtime.mount({
-        type: stagedFile.type,
-        source: stagedFile.source,
-        target: stagedFile.target,
-        mode: "readwrite",
-        metadata: stagedFile.metadata,
-      }))
-      interruption?.throwIfInterrupted()
-    }
-
-    const muPluginInstallCode = installMuPluginsCode(extraPlugins)
-    if (muPluginInstallCode) {
-      executions.push(withRecipeExecutionPhase(await runtime.execute({ command: "wordpress.run-php", args: [`code=${muPluginInstallCode}`] }), "setup", -2, "extra-plugin.install-mu-loader"))
-    }
-
-    const activatedPlugins = extraPlugins.filter((plugin) => plugin.loadAs === "plugin" && plugin.activate !== false)
-    if (activatedPlugins.length > 0) {
-      const activePluginsAfterActivation = await phaseTracker.run("activate_plugins", phasePluginActivationData(activatedPlugins), async () => {
-        for (const plugin of activatedPlugins) {
-          executions.push(withRecipeExecutionPhase(await runtime!.execute({ command: "wordpress.run-php", args: [`code=${activateExtraPluginCode(plugin.pluginFile)}`] }), "setup", -1, `extra-plugin.activate:${plugin.pluginFile}`))
-          interruption?.throwIfInterrupted()
-        }
-        return await activePlugins(runtime!)
-      })
-      const activationPhase = [...phaseTracker.list()].reverse().find((phase: RecipePhaseEvidence) => phase.name === "activate_plugins")
-      if (activationPhase?.data) {
-        activationPhase.data.activePlugins = activePluginsAfterActivation
-      }
-    }
-
-    for (const [index, setupStep] of (recipe.inputs?.pluginRuntime?.setup ?? []).entries()) {
-      executions.push(await awaitRecipe(`plugin-runtime.setup[${index}]`, executeRecipePluginRuntimeStep(runtime, setupStep, recipeDirectory, "setup", index)))
-      interruption?.throwIfInterrupted()
-    }
-
-    for (const [index, probe] of (recipe.inputs?.pluginRuntime?.healthProbes ?? []).entries()) {
-      executions.push(await awaitRecipe(`plugin-runtime.health:${probe.name}`, executeRecipePluginRuntimeHealthProbe(runtime, probe, recipeDirectory, index)))
-      interruption?.throwIfInterrupted()
-    }
+    executions.push(...(await applyRecipeRuntimeSetup({ recipe, recipeDirectory, runtime, prepared: preparedRuntimeSetup, phaseExecutor, interruption })).executions)
 
     fixtureDatabases = await phaseTracker.run("import_fixture_databases", phaseFixtureDatabaseData(recipe), async () => await awaitRecipe("fixture-databases.import", importRecipeFixtureDatabases(recipe, recipeDirectory, runtime!, executions)))
     const siteSeeds = await awaitRecipe("site-seeds.import", importRecipeSiteSeeds(recipe, recipeDirectory, runtime!, executions))
@@ -803,15 +681,6 @@ function classifyRecipePhaseFailure(phase: string): string {
   }
 }
 
-async function prepareRecipeRuntimeOverlaysForRun(recipe: WorkspaceRecipe, recipeDirectory: string): Promise<PreparedRuntimeOverlay[]> {
-  try {
-    return await prepareRecipeRuntimeOverlays(recipe, recipeDirectory)
-  } catch (error) {
-    const message = error instanceof Error ? error.message : String(error)
-    throw new Error(`Recipe runtime overlay preparation failed: ${message}`, { cause: error })
-  }
-}
-
 async function validateRecipe(options: RecipeValidateOptions): Promise<RecipeValidateOutput> {
   const recipePath = resolve(options.recipePath)
   try {
@@ -1109,26 +978,6 @@ async function executeRecipeWorkflowStep(runtime: Runtime, workflowStep: ReturnT
   } catch (error) {
     const message = error instanceof Error ? error.message : String(error)
     throw new Error(`Recipe workflow ${workflowStep.phase}[${workflowStep.index}] failed: ${message}`, { cause: error })
-  }
-}
-
-async function executeRecipePluginRuntimeStep(runtime: Runtime, step: WorkspaceRecipe["workflow"]["steps"][number], recipeDirectory: string, phase: "setup", index: number): Promise<RecipeExecutionResult> {
-  try {
-    const execution = await runtime.execute(await recipeExecutionSpec(step, recipeDirectory))
-    return withRecipeExecutionPhase(execution, phase, pluginRuntimeSetupStepIndex(index), `plugin-runtime.setup:${index}`)
-  } catch (error) {
-    const message = error instanceof Error ? error.message : String(error)
-    throw new Error(`Recipe plugin runtime setup[${index}] failed: ${message}`, { cause: error })
-  }
-}
-
-async function executeRecipePluginRuntimeHealthProbe(runtime: Runtime, probe: WorkspaceRecipePluginRuntimeHealthProbe, recipeDirectory: string, index: number): Promise<RecipeExecutionResult> {
-  try {
-    const execution = await runtime.execute(await recipeExecutionSpec(pluginRuntimeHealthProbeStep(probe), recipeDirectory))
-    return withRecipeExecutionPhase(execution, "setup", pluginRuntimeHealthProbeStepIndex(index), `plugin-runtime.health:${probe.name}`)
-  } catch (error) {
-    const message = error instanceof Error ? error.message : String(error)
-    throw new Error(`Recipe plugin runtime health probe "${probe.name}" failed: ${message}`, { cause: error })
   }
 }
 
@@ -1614,170 +1463,6 @@ if (is_file($path)) {
 echo wp_json_encode($result);`
 }
 
-function activateExtraPluginCode(pluginFile: string): string {
-  return `${runtimeComponentLifecycleReplayPhp("wp_codebox_activate_plugin")}
-$plugin_file = ${JSON.stringify(pluginFile)};
-require_once ABSPATH . 'wp-admin/includes/plugin.php';
-if (is_plugin_active($plugin_file)) {
-    deactivate_plugins($plugin_file, true, false);
-}
-$lifecycle = wp_codebox_activate_plugin_component_lifecycle_replay_prepare();
-$result = activate_plugin($plugin_file, '', false, false);
-wp_codebox_activate_plugin_component_lifecycle_replay_complete($lifecycle);
-if (is_wp_error($result)) {
-    throw new RuntimeException('Failed to activate extra plugin ' . $plugin_file . ': ' . $result->get_error_message());
-}
-do_action('wp_codebox_runtime_plugin_activated', $plugin_file);
-echo wp_json_encode(array('activated' => $plugin_file));`
-}
-
-function runtimeComponentLifecycleReplayPhp(prefix: string): string {
-  const snapshot = `${prefix}_component_lifecycle_snapshot_hook_callbacks`
-  const defer = `${prefix}_component_lifecycle_defer_new_hook_callbacks`
-  const run = `${prefix}_component_lifecycle_run_deferred_hook_callbacks`
-  const reopen = `${prefix}_component_lifecycle_reopen_action`
-  const restore = `${prefix}_component_lifecycle_restore_action`
-  const abilityNames = `${prefix}_component_lifecycle_ability_names`
-  const prepare = `${prefix}_component_lifecycle_replay_prepare`
-  const complete = `${prefix}_component_lifecycle_replay_complete`
-
-  return `if (!function_exists('${prepare}')) {
-function ${snapshot}(string $hook_name): array {
-    global $wp_filter;
-    $snapshot = array();
-    if (!isset($wp_filter[$hook_name]) || !isset($wp_filter[$hook_name]->callbacks)) {
-        return $snapshot;
-    }
-    foreach ($wp_filter[$hook_name]->callbacks as $priority => $callbacks) {
-        foreach (array_keys($callbacks) as $callback_id) {
-            $snapshot[$priority . ':' . $callback_id] = true;
-        }
-    }
-    return $snapshot;
-}
-function ${defer}(string $hook_name, array $before): array {
-    global $wp_filter;
-    $deferred = array();
-    if (!isset($wp_filter[$hook_name]) || !isset($wp_filter[$hook_name]->callbacks)) {
-        return $deferred;
-    }
-    foreach ($wp_filter[$hook_name]->callbacks as $priority => $callbacks) {
-        foreach ($callbacks as $callback_id => $callback) {
-            if (isset($before[$priority . ':' . $callback_id])) {
-                continue;
-            }
-            $deferred[] = array('priority' => (int) $priority, 'callback' => $callback);
-            unset($wp_filter[$hook_name]->callbacks[$priority][$callback_id]);
-        }
-        if (empty($wp_filter[$hook_name]->callbacks[$priority])) {
-            unset($wp_filter[$hook_name]->callbacks[$priority]);
-        }
-    }
-    usort($deferred, static function (array $left, array $right): int { return ($left['priority'] ?? 10) <=> ($right['priority'] ?? 10); });
-    return $deferred;
-}
-function ${run}(array $deferred, string $hook_name, array $args = array()): void {
-    global $wp_current_filter;
-    if (!is_array($wp_current_filter)) {
-        $wp_current_filter = array();
-    }
-    $wp_current_filter[] = $hook_name;
-    try {
-        foreach ($deferred as $entry) {
-            $callback = $entry['callback'] ?? null;
-            if (!is_array($callback) || !isset($callback['function'])) {
-                continue;
-            }
-            $accepted_args = isset($callback['accepted_args']) ? (int) $callback['accepted_args'] : count($args);
-            call_user_func_array($callback['function'], array_slice($args, 0, $accepted_args));
-        }
-    } finally {
-        array_pop($wp_current_filter);
-    }
-}
-function ${reopen}(string $hook_name): int {
-    global $wp_actions;
-    $count = function_exists('did_action') ? (int) did_action($hook_name) : 0;
-    if ($count > 0) {
-        if (!is_array($wp_actions)) {
-            $wp_actions = array();
-        }
-        $wp_actions[$hook_name] = 0;
-    }
-    return $count;
-}
-function ${restore}(string $hook_name, int $count): void {
-    global $wp_actions;
-    if ($count <= 0) {
-        return;
-    }
-    if (!is_array($wp_actions)) {
-        $wp_actions = array();
-    }
-    $wp_actions[$hook_name] = max($count, (int) ($wp_actions[$hook_name] ?? 0), 1);
-}
-function ${abilityNames}(): array {
-    if (!function_exists('wp_get_abilities')) {
-        return array();
-    }
-    $abilities = wp_get_abilities();
-    return is_array($abilities) ? array_values(array_map('strval', array_keys($abilities))) : array();
-}
-function ${prepare}(): array {
-    $hooks = array('plugins_loaded', 'init', 'wp_abilities_api_categories_init', 'wp_abilities_api_init', 'wp_codebox_runtime_abilities_ready');
-    $state = array('hooks' => array(), 'abilities_before' => ${abilityNames}());
-    foreach ($hooks as $hook_name) {
-        $state['hooks'][$hook_name] = array(
-            'callbacks' => ${snapshot}($hook_name),
-            'did_action' => ${reopen}($hook_name),
-        );
-    }
-    return $state;
-}
-function ${complete}(array $state): array {
-    $diagnostic = array(
-        'schema' => 'wp-codebox/runtime-component-lifecycle-replay/v1',
-        'hooks' => array(),
-        'abilities_before' => $state['abilities_before'] ?? array(),
-        'abilities_after' => array(),
-        'abilities_added' => array(),
-    );
-    foreach (($state['hooks'] ?? array()) as $hook_name => $hook_state) {
-        $deferred = ${defer}((string) $hook_name, is_array($hook_state['callbacks'] ?? null) ? $hook_state['callbacks'] : array());
-        ${run}($deferred, (string) $hook_name);
-        ${restore}((string) $hook_name, (int) ($hook_state['did_action'] ?? 0));
-        $diagnostic['hooks'][(string) $hook_name] = array('replayed_callbacks' => count($deferred), 'previous_did_action' => (int) ($hook_state['did_action'] ?? 0));
-    }
-    $diagnostic['abilities_after'] = ${abilityNames}();
-    $diagnostic['abilities_added'] = array_values(array_diff($diagnostic['abilities_after'], is_array($diagnostic['abilities_before']) ? $diagnostic['abilities_before'] : array()));
-    return $diagnostic;
-}
-}`
-}
-
-async function activePlugins(runtime: Runtime): Promise<string[]> {
-  const execution = await runtime.execute({
-    command: "wordpress.run-php",
-    args: ["code=echo wp_json_encode(array_values((array) get_option('active_plugins', array())));"],
-  })
-  const parsed = JSON.parse(execution.stdout.trim() || "[]") as unknown
-  return Array.isArray(parsed) ? parsed.filter((plugin): plugin is string => typeof plugin === "string") : []
-}
-
-function phasePluginMountData(extraPlugins: PreparedExtraPlugin[]): Record<string, unknown> {
-  return {
-    count: extraPlugins.length,
-    plugins: extraPlugins.map((plugin) => ({ slug: plugin.slug, pluginFile: plugin.pluginFile, target: plugin.target, loadAs: plugin.loadAs })),
-  }
-}
-
-function phasePluginActivationData(activatedPlugins: PreparedExtraPlugin[]): Record<string, unknown> {
-  return {
-    count: activatedPlugins.length,
-    plugins: activatedPlugins.map((plugin) => ({ slug: plugin.slug, pluginFile: plugin.pluginFile })),
-  }
-}
-
 function phaseWorkflowData(workflowSteps: ReturnType<typeof recipeWorkflowSteps>): Record<string, unknown> {
   return {
     count: workflowSteps.length,
@@ -2201,72 +1886,6 @@ function activePluginPhaseIncludes(phases: RecipePhaseEvidence[], pluginFile: st
   return phases.some((phase) => phase.name === "activate_plugins" && Array.isArray(phase.data?.activePlugins) && phase.data.activePlugins.includes(pluginFile))
 }
 
-async function inputMountMetadataWithBaseline(source: string, mount: WorkspaceRecipeMount, cleanupPaths: string[]): Promise<Record<string, unknown> | undefined> {
-  const metadata = mount.metadata ? { ...mount.metadata } : {}
-  if ((mount.mode ?? "readwrite") !== "readwrite") {
-    return Object.keys(metadata).length > 0 ? metadata : undefined
-  }
-  if (typeof metadata.baselineSource === "string" && metadata.baselineSource.length > 0) {
-    return metadata
-  }
-
-  let sourceStats
-  try {
-    sourceStats = await stat(source)
-  } catch {
-    return Object.keys(metadata).length > 0 ? metadata : undefined
-  }
-  if (!sourceStats.isDirectory() || await hasGitMetadata(source)) {
-    return Object.keys(metadata).length > 0 ? metadata : undefined
-  }
-
-  const baselineSource = await mkdtemp(join(tmpdir(), "wp-codebox-input-mount-baseline-"))
-  cleanupPaths.push(baselineSource)
-  await cp(source, baselineSource, {
-    recursive: true,
-    filter: (entry) => shouldCopyInputMountBaselineEntry(source, entry),
-  })
-  metadata.baselineSource = baselineSource
-  metadata.baselineStrategy = "input-mount-snapshot"
-  return metadata
-}
-
-async function cleanupInputMountBaselines(paths: string[]): Promise<void> {
-  await Promise.all(paths.map((path) => rm(path, { recursive: true, force: true })))
-  paths.length = 0
-}
-
-async function hasGitMetadata(directory: string): Promise<boolean> {
-  try {
-    await stat(join(directory, ".git"))
-    return true
-  } catch {
-    return false
-  }
-}
-
-function shouldCopyInputMountBaselineEntry(sourceRoot: string, entry: string): boolean {
-  const relativePath = entry.slice(sourceRoot.length).replace(/^\/+/, "")
-  if (!relativePath) {
-    return true
-  }
-  const firstSegment = relativePath.split("/")[0]
-  return firstSegment !== ".git" && firstSegment !== "node_modules"
-}
-
-function recipeRunDependencyOverlay(overlay: PreparedDependencyOverlay): Record<string, unknown> {
-  return {
-    source: overlay.source,
-    sourceRef: overlay.sourceRef,
-    target: overlay.target,
-    package: overlay.package,
-    consumer: overlay.consumer,
-    type: overlay.type,
-    mode: overlay.mode,
-    metadata: overlay.metadata,
-  }
-}
-
 function previewCliOverrides(recipePreview: RuntimePreviewSpec | undefined, effectivePreview: RuntimePreviewSpec): RuntimePreviewSpec | undefined {
   const overrides = stripUndefined({
     publicUrl: recipePreview?.publicUrl !== effectivePreview.publicUrl ? effectivePreview.publicUrl : undefined,
@@ -2282,32 +1901,6 @@ function previewMetadataSource(recipePreview: RuntimePreviewSpec | undefined, ef
   }
 
   return recipePreview ? "recipe-runtime-preview" : "cli-preview-options"
-}
-
-function recipeRunStagedFile(stagedFile: PreparedStagedFile): RecipeRunStagedFile {
-  const index = typeof stagedFile.metadata.index === "number" ? stagedFile.metadata.index : 0
-  return {
-    index,
-    source: stagedFile.originalSource,
-    sourceRef: stagedFile.sourceRef,
-    target: stagedFile.target,
-    type: stagedFile.type,
-    provenance: stagedFile.provenance,
-    action: "staged",
-  }
-}
-
-function recipeRunExtraPlugin(plugin: PreparedExtraPlugin): Record<string, unknown> {
-  return {
-    source: plugin.source,
-    slug: plugin.slug,
-    target: plugin.target,
-    pluginFile: plugin.pluginFile,
-    activate: plugin.activate,
-    loadAs: plugin.loadAs,
-    provenance: plugin.provenance,
-    metadata: plugin.metadata,
-  }
 }
 
 async function importRecipeSiteSeeds(recipe: WorkspaceRecipe, recipeDirectory: string, runtime: Runtime, executions: RecipeExecutionResult[]): Promise<RecipeRunSiteSeed[]> {
