+++ scripts/runtime-component-lifecycle-replay-smoke.ts
@@ -37,8 +37,8 @@ assert.ok(
   "late include bootstrap should replay newly registered callbacks after including plugin code",
 )
 
-const recipeRunSource = readFileSync(join(process.cwd(), "packages/cli/src/commands/recipe-run.ts"), "utf8")
-const activateFunction = recipeRunSource.slice(recipeRunSource.indexOf("function activateExtraPluginCode"), recipeRunSource.indexOf("async function activePlugins"))
+const recipeRuntimeSetupSource = readFileSync(join(process.cwd(), "packages/cli/src/commands/recipe-runtime-setup.ts"), "utf8")
+const activateFunction = recipeRuntimeSetupSource.slice(recipeRuntimeSetupSource.indexOf("function activateExtraPluginCode"), recipeRuntimeSetupSource.length)
 
 assert.match(activateFunction, /wp-codebox\/runtime-component-lifecycle-replay\/v1/, "activation setup should expose lifecycle replay diagnostics")
 assert.match(activateFunction, /wp_codebox_runtime_abilities_ready/, "activation setup should replay the post-abilities contract hook")
